NMS-14047: Refactoring
