+++ features/device-config/rest/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -4,7 +4,6 @@
 >
     <reference id="deviceConfigDao" interface="org.opennms.features.deviceconfig.persistence.api.DeviceConfigDao" availability="mandatory"/>
     <reference id="monitoredServiceDao" interface="org.opennms.netmgt.dao.api.MonitoredServiceDao" availability="mandatory"/>
-
     <reference id="deviceConfigService" interface="org.opennms.features.deviceconfig.service.DeviceConfigService" availability="mandatory"/>
 
     <bean id="defaultDeviceConfigRestService"
